pyproject.toml
[project]
name = "care-pd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9.13"
dependencies = [
    "torch-geometric>=2.6.1",
    "numpy<2.0.0",
    "greenlet<3.2",
    "einops==0.8.1",
    "imageio==2.31.5",
    "joblib==1.4.2",
    "matplotlib==3.8.0",
    "numpyencoder==0.3.0",
    "optuna==4.2.1",
    "pandas==2.2.3",
    "scikit-learn==1.6.1",
    "scipy>=1.13.1",
    "seaborn>=0.13.2",
    "setuptools==75.8.0",
    "smplx==0.1.28",
    "timm==1.0.15",
    "torch-dct==0.1.6",
    "tqdm==4.66.1",
    "trimesh==4.6.8",
    "wandb>=0.25.0",
    "gdown>=5.2.1",
    "chumpy>=0.70",
]

[tool.uv.extra-build-dependencies]
chumpy = ["pip"]
